Disable CMake language support to speed up Hunter download.
HunterGate.cmake creates a CMakeLists.txt just to download hunter.  It doesn't compile anything and doesn't need programming language support.  The benefit of disabling language support is that CMake doesn't try to initialise and test the compilers, which saves time and avoids printing those steps in the output.  They were quite confusing because, when a non-default generator is chosen, the download step still used the default generator and made it seem like the wrong compiler was being used, even though that compiler wasn't used at all.

This change explicitly declares the project with no language support.